Join Investigators Listed in Investigation Schedule & Case Assignment Project

asg_join_investigators() joins the investigators scheduled to work on date with the list of investigators from the Case Assignment REDcap project. By default, it performs an inner join, since only investigators in both lists are available for case assignment.

asg_join_investigators(
  date = Sys.Date(),
  type = c("inner", "anti_schedule", "anti_redcap"),
  api_token = Sys.getenv("redcap_CA_token"),
  quiet = FALSE
)

Arguments

date
The date to use when checking whether investigators are scheduled

type
The type of join to use. "inner" returns only investigators in both lists (the default). "anti_join_scheduled" returns investigators who are scheduled but not listed in REDcap; "anti_join_redcap" returns investigators who are listed in REDcap but not scheduled.

api_token
The API token for the Case Assignment REDcap project. The default pulls this from the environment variable redcap_CA_token.

quiet
Should update messages be suppressed?

Value

A tibble with one row per investigator and an investigator column. If type = "inner", an id column is also included.

Details

While the default behavior is an inner join, asg_join_investigators() can also perform anti-joins with either the scheduled or REDcap investigator list as the primary table. This is useful for debugging and checking that all scheduled investigators are able to have cases assigned.