Validate Case Assignment Upload

validate_upload() validates that all assigned cases were uploaded. By default, it archives accepted and rejected cases and notifies relevant parties of the status. Note that, unlike validate_assignments(), a notification is sent even if all cases are uploaded correctly.

validate_upload(
  .data,
  archive = TRUE,
  notify_to = c("Jesse.Smith@shelbycountytn.gov", "Karim.Gilani@shelbycountytn.gov"),
  cnd = c("warning", "error"),
  date = attr(.data, "date")
)

Arguments

.data
Upload data from upload_assigments()

archive
Should accepted and rejected cases be archived?

notify_to
Email addresses to notify of case assignment. Set to NULL to disable notification.

cnd
Should rejected cases trigger a warning or an error?

date
The date to use in notification

Value

The input data